requirements.txt
streamlit
langchain
langchain-classic
langchain-openai
langchain-community
chromadb
pypdf
python-dotenv
tiktoken